Extracting `#[before_fund]` and  `#[fundable]` specific code from `function.rs` into separate files.
